Cellpose train preprocessing development
To train cellopse we need to make sure data in correct form.
Therefore, wrote code to load back in parquet dataframe and render the
histogram again using the x_pixel, y_pixel values using quick polars
method
Need to write rest so takes in parquet file with gt annotations and
creates the img and segmentation for training

Further, I changed the operation of coord_2_histo so that the last
localisation is not manually moved to the last pixel.

Instead the bins are increased by 0.1% so that the last localisation
fits in the last bin (i.e. all bins stretched by same ratio)

[Ticket: X]